+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Features:
   + Structured documentation comments for are now supported for function definitions.
+  + Auto-generated struct constructors can now be documented.
 - Bug fixes:
   + Links to members from documentation comments of function parameters, return type, and throws
     clause now work correctly in C++, Java, and Swift.